feat: refactor authentication setup and middleware integration
