+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,51 +1,6 @@
-import NextAuth, { AuthOptions } from "next-auth";
-import CredentialsProvider from "next-auth/providers/credentials";
-import prisma from "@/lib/prisma";
-import bcrypt from "bcryptjs";
-
-export const authOptions: AuthOptions = {
-  providers: [
-    CredentialsProvider({
-      name: "Credentials",
-      credentials: {
-        email: { label: "Email", type: "text" },
-        password: { label: "Password", type: "password" },
-      },
-      async authorize(credentials) {
-        if (!credentials?.email || !credentials?.password) return null;
-        const user = await prisma.user.findUnique({
-          where: { email: credentials.email },
-        });
-
-        if (!user) {
-          throw new Error("Invalid email or password");
-        }
-
-        const isValid = await bcrypt.compare(
-          credentials.password,
-          user.password
-        );
-
-        if (!isValid) {
-          throw new Error("Invalid email or password");
-        }
-
-        const { password, ...userWithoutPassword } = user; // Exclude password from the user object
-        return userWithoutPassword;
-      },
-    }),
-  ],
-  pages: {
-    signIn: "/auth/signin", // Redirect unauthenticated users here
-    error: "/auth/signin", // Error code passed in query string as ?error=
-  },
-  session: {
-    strategy: "jwt",
-  },
-  secret: process.env.NEXTAUTH_SECRET,
-};
+import NextAuth from "next-auth";
+import authOptions from "@lib/auth";
 
 const handler = NextAuth(authOptions);
-export const auth = handler.auth;
 
 export { handler as GET, handler as POST };
