+++ src/components/Providers.tsx
@@ -2,6 +2,6 @@
 import { SessionProvider } from "next-auth/react";
 import React from "react";
 
-export default function ({ children }: React.PropsWithChildren) {
+export default function Providers({ children }: React.PropsWithChildren) {
   return <SessionProvider>{children}</SessionProvider>;
 }
